Improve CRLF handling for Windows
Current implementation assumes the input integers are always separated by a single character. However, this is not true on Windows due to the use or CR (0x0D) in conjunction with LF (0x0A) for line separation. This causes segmentation faults. We fix this by simply skipping characters <= 0x20 on every integer/float/word input.